Port trends command to new command system
